Fix comparison between size_t and unsigned long
Because size_t is not always the same type as unsigned long. Size_t can vary with how large the address size is (though we only support 64 bit now) and unsigned long can vary with whatever it thinks the best word size is for integers.

+++ tests/arcus/ArcusCommunicationPrivateTest.cpp
@@ -195,15 +195,15 @@ void ArcusCommunicationPrivateTest::readMeshGroupMessageTest()
 
     // Checks:
     auto& scene = Application::getInstance().current_slice->scene;
-    CPPUNIT_ASSERT(! scene.mesh_groups.empty());
+    CPPUNIT_ASSERT(!scene.mesh_groups.empty());
 
     auto& meshes = scene.mesh_groups[0].meshes;
-    CPPUNIT_ASSERT(! meshes.empty());
+    CPPUNIT_ASSERT(!meshes.empty());
 
     auto& vertices = meshes[0].vertices;
-    CPPUNIT_ASSERT(! vertices.empty());
-    CPPUNIT_ASSERT_EQUAL(vertices.size(), 8ul); //A cube should have 8 unique vertices.
-    CPPUNIT_ASSERT_EQUAL(meshes[0].faces.size(), 12ul); // A cube should have 12 tri-s (2 for each 6 sides of the dice).
+    CPPUNIT_ASSERT(!vertices.empty());
+    CPPUNIT_ASSERT_EQUAL(vertices.size(), size_t(8)); //A cube should have 8 unique vertices.
+    CPPUNIT_ASSERT_EQUAL(meshes[0].faces.size(), size_t(12)); // A cube should have 12 tri-s (2 for each 6 sides of the dice).
 
     // Distances should be the same:
